Parse the new PrivCount events in the data collector and injector
Update test_tor_ctl_port.py for the new event argument format.
Update the integration tests for the new event format.

This restores PrivCount compatibility with the tor patch in #99, but only
by activating all the events, all the time.

To achieve the efficiency goals of #100, we must choose the events passed
to SETEVENTS based on the counters in each round.